ESQL: Move expression translators
This moves `ExpressionTranslator` implementations that we use from
esql-core into esql. It moves the classes into their own file and
deletes the translators we don't currently use.
